Fix bug on 64-bit systems (NSNotFound comparison with 32-bit value).
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileFoundation@7916 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Source/NSIndexSet+Etoile.m
@@ -18,7 +18,7 @@ Each resulting index path only contains a single index. */
 {
 	NSMutableArray *indexPaths = [NSMutableArray arrayWithCapacity: [self count]];
 	/* Will set lastIndex to 0 or NSNotFound */
-	unsigned int lastIndex = [self indexGreaterThanOrEqualToIndex: 0];
+	NSUInteger lastIndex = [self indexGreaterThanOrEqualToIndex: 0];
 
 	if (lastIndex == NSNotFound)
 		return nil;
